+++ marinechoice/lib/pages/mappage.dart
@@ -1,12 +1,12 @@
+import 'package:firebase_database/firebase_database.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:location/location.dart';
-import 'package:marinechoice/models/fisharea_model.dart';
 import 'package:marinechoice/pages/recipespage.dart';
-
-import '../dbhelper.dart';
+import '../models/fisharea_model.dart';
 import '../models/protarea_model.dart';
+import 'SettingsPage.dart';
 import 'homepage.dart';
 
 class MapPage extends StatefulWidget {
@@ -19,47 +19,43 @@ class MapPage extends StatefulWidget {
 class _MapPageState extends State<MapPage> {
   Location _locationController = new Location();
   LatLng? _currentP = null;
+  final _database = FirebaseDatabase.instance.ref();
   Future<Set<Marker>?>? _pointsFuture;
+
+  List<ProtArea> protAreaList = [];
+  List<FishArea> fisAreaList = [];
 
   @override
   void initState() {
     super.initState();
     _pointsFuture = getPoints();
     getLocationUpdates();
-
   }
 
   Future<Set<Marker>?> getPoints() async {
+
     try {
       Set<Marker> set = {};
-      var protectedAreas = await DataBaseHelperFilters.getAllProtectedAreas();
+      await retrieveMapData();
 
-      if (protectedAreas == null) {
-        return set;
-      }
 
-      for (var element in protectedAreas) {
+
+      for (var element in protAreaList) {
         set.add(Marker(
-            infoWindow: InfoWindow(title: "Area Protegida",snippet: element.description),
-            markerId: MarkerId("p_area_${element.idAreaP}"),
+            infoWindow: InfoWindow(title: "Area Protegida",snippet: element.protAreaData!.description!),
+            markerId: MarkerId("p_area_${element.protAreaData!.description!}"),
             icon:
             BitmapDescriptor.defaultMarkerWithHue(BitmapDescriptor.hueOrange),
-            position: LatLng(element.latitude, element.longitude)));
+            position: LatLng(double.parse(element.protAreaData!.latitude!), double.parse(element.protAreaData!.longitude!))));
       }
 
-      var fishAreas = await DataBaseHelperFilters.getAllFishAreas();
-
-      if (fishAreas == null) {
-        return set;
-      }
-
-      for (var element in fishAreas) {
+      for (var element in fisAreaList) {
         set.add(Marker(
-            infoWindow: InfoWindow(title: "Area de Pesca",snippet: "Pesca de ${element.fish.name}"),
-            markerId: MarkerId("f_area_${element.idAreaF}"),
+            infoWindow: InfoWindow(title: "Area de Pesca",snippet: "Pesca de ${element.fishAreaData!.fish!}"),
+            markerId: MarkerId("f_area_${element.fishAreaData!.fish!}"),
             icon: BitmapDescriptor.defaultMarkerWithHue(
                 BitmapDescriptor.hueMagenta),
-            position: LatLng(element.latitude, element.longitude)));
+            position: LatLng(double.parse(element.fishAreaData!.latitude!), double.parse(element.fishAreaData!.longitude!))));
       }
 
 
@@ -193,7 +189,7 @@ class _MapPageState extends State<MapPage> {
             height: 25,
             width: 30,
           ),
-          label: ("FIND"),
+          label: ("FISH"),
         ),
         BottomNavigationBarItem(
             icon: SvgPicture.asset(
@@ -215,7 +211,7 @@ class _MapPageState extends State<MapPage> {
               height: 25,
               width: 30,
             ),
-            label: ("SHARE")),
+            label: ("POST")),
         BottomNavigationBarItem(
             icon: SvgPicture.asset(
               'assets/icons/user.svg',
@@ -260,7 +256,10 @@ class _MapPageState extends State<MapPage> {
       backgroundColor: const Color(0xffB4D8F9),
       actions: [
         GestureDetector(
-          onTap: () {},
+          onTap: () {
+            Navigator.of(context).push(
+                MaterialPageRoute(builder: (context) => const SettingsPage()));
+          },
           child: Container(
             margin: const EdgeInsets.all(10),
             alignment: Alignment.center,
@@ -277,5 +276,29 @@ class _MapPageState extends State<MapPage> {
         )
       ],
     );
+  }
+
+  Future<void> retrieveMapData() async {
+
+    var fishAreas = await _database.child("FISHAREA").get();
+
+    for(var fishArea in fishAreas.children){
+      FishAreaData fishAreaData = FishAreaData.fromJson(fishArea.value as Map);
+      FishArea fishAreaF = FishArea(key: fishArea.key, fishAreaData: fishAreaData);
+      fisAreaList.add(fishAreaF);
+    }
+
+    print("List $fisAreaList");
+
+    var protAreas = await _database.child("PROTAREA").get();
+
+    for(var protArea in protAreas.children){
+      ProtAreaData protAreaData = ProtAreaData.fromJson(protArea.value as Map);
+      ProtArea protAreaf = ProtArea(key: protArea.key, protAreaData: protAreaData);
+      protAreaList.add(protAreaf);
+    }
+
+    print("List $protAreaList");
+
   }
 }