+++ marinechoice/lib/pages/recipespage.dart
@@ -1,37 +1,36 @@
-
+import 'package:firebase_database/firebase_database.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_svg/svg.dart';
-import 'package:marinechoice/dbhelper.dart';
 import 'package:marinechoice/pages/recipeinfopage.dart';
 
 import '../models/recipe_model.dart';
+import 'SettingsPage.dart';
 import 'homepage.dart';
 import 'mappage.dart';
 
-
-class RecipesPage extends StatefulWidget{
+class RecipesPage extends StatefulWidget {
   const RecipesPage({super.key});
 
   @override
   State<StatefulWidget> createState() {
     return _RecipesPage();
   }
-
-
-
 }
 
 class _RecipesPage extends State<RecipesPage> {
+  final _database = FirebaseDatabase.instance.ref();
+
+  List<Recipe> recipeList = [];
+
+  @override
+  void initState() {
+    super.initState();
+  }
 
   @override
   Widget build(BuildContext context) {
-    final titleController = TextEditingController();
-    final preparationController = TextEditingController();
-    final authorController = TextEditingController();
-
-
     return Scaffold(
       backgroundColor: const Color(0xffB4D8F9),
       appBar: buildAppBar(),
@@ -41,92 +40,93 @@ class _RecipesPage extends State<RecipesPage> {
   }
 
   Widget buildContainer(Recipe recipe, int number) {
-      return GestureDetector(
-        onTap: () =>{Navigator.of(context).push(MaterialPageRoute(builder: (context) => RecipeInfoPage(recipe:recipe)))},
-        child:Container(
+    if (recipeList.isEmpty) {
+      return Center(
+        child: CircularProgressIndicator(),
+      );
+    }
+
+    return GestureDetector(
+      onTap: () => {
+        Navigator.of(context).push(MaterialPageRoute(
+            builder: (context) => RecipeInfoPage(recipe: recipe)))
+      },
+      child: Container(
         padding: const EdgeInsets.all(20),
-        width: MediaQuery.of(context).size.width/2 - 20,
+        width: MediaQuery.of(context).size.width / 2 - 20,
         height: 130,
         decoration: BoxDecoration(
             border: Border.all(color: const Color(0xffD6E7F7), width: 3),
             color: const Color(0xffD6E7F7),
             borderRadius: BorderRadius.circular(15)),
         margin: const EdgeInsets.all(10),
-        child: Column(
-          children:[ Text(
-            recipe.title,
+        child: Column(children: [
+          Text(
+            recipe.recipeData!.title!,
             style: const TextStyle(
               fontSize: 15,
               fontWeight: FontWeight.w400,
             ),
-          ),]
-        ),
-      ),);
-
+          ),
+        ]),
+      ),
+    );
   }
 
   Widget buildSingleChildScrollView() {
-
-    return FutureBuilder<List<Recipe>?>(
-        future: DataBaseHelperFilters.getAllRecipes(),
-        builder: (context,snapshot){
-          if (snapshot.connectionState == ConnectionState.done){
-
-            print(snapshot.data);
+    return FutureBuilder<void>(
+        future: retrieveRecipeData(),
+        builder: (context, snapshot) {
+          if (snapshot.connectionState == ConnectionState.done) {
             return SingleChildScrollView(
-              child: Column(
-                children: [
-                  Container(
-                    margin: const EdgeInsets.all(30),
-                    child: const Text("Recommended",
-                    textAlign: TextAlign.left,
-                    style: TextStyle(
-                      fontSize: 32,
-                      fontWeight: FontWeight.w700,
-                    )),
-                  ),
-                  Wrap(
+              child: Column(children: [
+                Container(
+                  margin: const EdgeInsets.all(30),
+                  child: const Text("Recommended",
+                      textAlign: TextAlign.left,
+                      style: TextStyle(
+                        fontSize: 32,
+                        fontWeight: FontWeight.w700,
+                      )),
+                ),
+                Wrap(
                     direction: Axis.horizontal,
-                  children: _getRecipes(snapshot.data)
-                ),]
-              ),
+                    children: _getRecipes(recipeList)),
+              ]),
             );
-          }else{
-            return const Center(child: CircularProgressIndicator());
+          } else {
+            return Center(
+              child: CircularProgressIndicator(),
+            );
           }
-
-
         });
-
-
   }
 
- List<Widget> _getRecipes(List<Recipe>? recipes) {
-    List<Widget>  list = [];
-    if (recipes == null){
-      return list;
-    }else{
-      for (int i = 0; i < recipes.length; i++) {list.add(buildContainer(recipes[i],i));}
-    }
-    return list;
- }
+  List<Widget> _getRecipes(List<Recipe> recipeList) {
+    List<Widget> list = [];
 
-  _navigate(int index){
+    for (int i = 0; i < recipeList.length; i++) {
+      list.add(buildContainer(recipeList[i], i));
+    }
+
+    return list;
+  }
+
+  _navigate(int index) {
     switch (index) {
       case 0:
-        Navigator.of(context).push(
-            MaterialPageRoute(builder: (context) => const HomePage()));
+        Navigator.of(context)
+            .push(MaterialPageRoute(builder: (context) => const HomePage()));
         break;
       case 2:
-        Navigator.of(context).push(
-            MaterialPageRoute(builder: (context) => const RecipesPage()));
+        Navigator.of(context)
+            .push(MaterialPageRoute(builder: (context) => const RecipesPage()));
         break;
       case 3:
-        Navigator.of(context).push(
-            MaterialPageRoute(builder: (context) => const MapPage()));
+        Navigator.of(context)
+            .push(MaterialPageRoute(builder: (context) => const MapPage()));
         break;
     }
-
   }
 
   BottomNavigationBar buildBottomNavigationBar() {
@@ -139,31 +139,55 @@ class _RecipesPage extends State<RecipesPage> {
       items: [
         BottomNavigationBarItem(
           icon: SvgPicture.asset(
-            'assets/icons/home.svg', height: 25, width: 30,),
-          label: ("HOME"),),
+            'assets/icons/home.svg',
+            height: 25,
+            width: 30,
+          ),
+          label: ("HOME"),
+        ),
         BottomNavigationBarItem(
           icon: SvgPicture.asset(
-            'assets/icons/fishing-rod.svg', height: 25, width: 30,),
-          label: ("FIND"),),
+            'assets/icons/fishing-rod.svg',
+            height: 25,
+            width: 30,
+          ),
+          label: ("FISH"),
+        ),
         BottomNavigationBarItem(
             icon: SvgPicture.asset(
-              'assets/icons/chef-hat.svg', height: 30, width: 30,),
+              'assets/icons/chef-hat.svg',
+              height: 30,
+              width: 30,
+            ),
             label: ("COOK")),
         BottomNavigationBarItem(
             icon: SvgPicture.asset(
-              'assets/icons/map.svg', height: 30, width: 30,),
+              'assets/icons/map.svg',
+              height: 30,
+              width: 30,
+            ),
             label: ("MAP")),
         BottomNavigationBarItem(
             icon: SvgPicture.asset(
-              'assets/icons/envelope-open.svg', height: 25, width: 30,),
-            label: ("SHARE")),
+              'assets/icons/envelope-open.svg',
+              height: 25,
+              width: 30,
+            ),
+            label: ("POST")),
         BottomNavigationBarItem(
             icon: SvgPicture.asset(
-              'assets/icons/user.svg', height: 30, width: 30,),
+              'assets/icons/user.svg',
+              height: 30,
+              width: 30,
+            ),
             label: ("YOU")),
       ],
-      selectedLabelStyle: const TextStyle(fontWeight: FontWeight.w900,),
-      unselectedLabelStyle: const TextStyle(fontWeight: FontWeight.w900,),
+      selectedLabelStyle: const TextStyle(
+        fontWeight: FontWeight.w900,
+      ),
+      unselectedLabelStyle: const TextStyle(
+        fontWeight: FontWeight.w900,
+      ),
     );
   }
 
@@ -184,7 +208,7 @@ class _RecipesPage extends State<RecipesPage> {
         decoration: InputDecoration(
           suffixIcon: Padding(
               padding:
-              const EdgeInsets.only(left: 20, top: 5, right: 0, bottom: 5),
+                  const EdgeInsets.only(left: 20, top: 5, right: 0, bottom: 5),
               child: SvgPicture.asset('assets/icons/search.svg')),
           hintText: 'Search...',
           hintStyle: TextStyle(color: Colors.black87),
@@ -194,7 +218,10 @@ class _RecipesPage extends State<RecipesPage> {
       backgroundColor: const Color(0xffB4D8F9),
       actions: [
         GestureDetector(
-          onTap: () {},
+          onTap: () {
+            Navigator.of(context).push(
+                MaterialPageRoute(builder: (context) => const SettingsPage()));
+          },
           child: Container(
             margin: const EdgeInsets.all(10),
             alignment: Alignment.center,
@@ -213,5 +240,13 @@ class _RecipesPage extends State<RecipesPage> {
     );
   }
 
+  Future<void> retrieveRecipeData() async {
+    var result = await _database.child("RECIPE").get();
 
+    for (var recipe in result.children) {
+      RecipeData recipeData = RecipeData.fromJson(recipe.value as Map);
+      Recipe recipef = Recipe(key: recipe.key, recipeData: recipeData);
+      recipeList.add(recipef);
+    }
+  }
 }