+++ changelog.md
@@ -3,7 +3,7 @@
 ## [1.0.0] - 21-04-2024
 
 - Changed some buttons design;
-- Change the search bar.
+- Changed the search bar.
 
 ## [1.0.1] - 25-04-2024
 
@@ -13,3 +13,15 @@
 - Created a settings page;
 - Created with firebase a way for the user to log out of the app;
 - Created with firebase a way to keep the user logged in.
+
+## [1.0.2] - 28-04-2024
+
+- Created a sever with Firebase (Realtime Database);
+- Recipe data is now stored and pulled from the server;
+- Fish Areas data is now stored and pulled from the server;
+- Protected Areas data is now stored and pulled from the server;
+- Settings page now is accessible through every page;
+- Changed name of some buttons;
+- Removed local database;
+- Removed all models related to local database;
+